+++ memo/src/compornents/main/Home.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+
+export const Home = () => {
+    return (
+        <div>
+            ホームです。
+        </div>
+    )
+}
